requirements.txt
soco==0.29.1
pillow==10.1.0
requests==2.31.0
